+++ pyccel/ast/internals.py
@@ -20,13 +20,13 @@ __all__ = (
     'PrecomputedCode',
     'PyccelArrayShapeElement',
     'PyccelArraySize',
-    'PyccelInternalFunction',
+    'PyccelFunction',
     'PyccelSymbol',
     'Slice',
 )
 
 
-class PyccelInternalFunction(TypedAstNode):
+class PyccelFunction(TypedAstNode):
     """
     Abstract class for function calls translated to Pyccel objects.
 
@@ -67,7 +67,7 @@ class PyccelInternalFunction(TypedAstNode):
         return False
 
 
-class PyccelArraySize(PyccelInternalFunction):
+class PyccelArraySize(PyccelFunction):
     """
     Gets the total number of elements in an array.
 
@@ -108,7 +108,7 @@ class PyccelArraySize(PyccelInternalFunction):
             return False
 
 
-class PyccelArrayShapeElement(PyccelInternalFunction):
+class PyccelArrayShapeElement(PyccelFunction):
     """
     Gets the number of elements in a given dimension of an array.