+++ pyccel/ast/itertoolsext.py
@@ -7,14 +7,14 @@ This module represent a call to the itertools functions for code generation.
 """
 from .builtins  import PythonLen, PythonRange
 from .core      import PyccelFunctionDef, Module
-from .internals import PyccelInternalFunction
+from .internals import PyccelFunction
 
 __all__ = (
     'Product',
     'itertools_mod',
 )
 
-class Product(PyccelInternalFunction):
+class Product(PyccelFunction):
     """
     Represents a call to itertools.product for code generation.